feat: re-write Beatstream in Golang and React
This is a major v2 release of Beatstream with backend writtend in Golang and the old jQuery UI quickly transformed to React. Overall, the app should be quite similar to before, though the playlist UI works a bit different, sorting is not as good, and ID3 tag parsing works better.